Fix critical runtime bugs and concurrency issues (BUG-0001 to BUG-0007)
- Fixed integer overflow in component merge (BUG-0001)
- Fixed stale pool bounds cache in handle system (BUG-0002)
- Fixed unsafe jmp_buf copy in frame_clone (BUG-0003)
- Fixed SCC refcount underflow (BUG-0004)
- Fixed symmetric pool type punning (BUG-0005)
- Fixed unchecked realloc in component (BUG-0006)
- Fixed potential strdup silent failure in constraint (BUG-0007)
- Fixed concurrency race in memory subsystem with mutexes and atomic RC
- Fixed weak reference registry memory leak
- Added runtime/tests/verify_fixes.c for regression testing

+++ runtime/src/memory/component.h
@@ -36,8 +36,8 @@ struct SymComponent {
     };
     
     struct SymObj** members; /* All objects in this island */
-    int member_count;
-    int member_capacity;
+    size_t member_count;
+    size_t member_capacity;
     
     struct SymComponent* parent; /* Union-Find parent (NULL if root) */
